requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
boto3==1.38.13; python_version >= '3.9'
botocore==1.38.13; python_version >= '3.9'
cffi==1.17.1; python_version >= '3.8'
click==8.2.0; python_version >= '3.10'
cryptography==44.0.3; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
fastapi==0.115.12; python_version >= '3.8'
h11==0.16.0; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
jmespath==1.0.1; python_version >= '3.7'
mangum==0.19.0; python_version >= '3.7'
pycparser==2.22; python_version >= '3.8'
pydantic==2.11.4; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pyjwt[crypto]==2.10.1; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
s3transfer==0.12.0; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.46.2; python_version >= '3.9'
typing-extensions==4.13.2; python_version >= '3.8'
typing-inspection==0.4.0; python_version >= '3.9'
urllib3==2.4.0; python_version >= '3.9'
uvicorn==0.34.2; python_version >= '3.9'
